自定义分表算法

+++ demo-sharding-sphere/src/main/java/com/songjy/sharding/sphere/model/order/Order.java
@@ -16,12 +16,12 @@ public class Order implements Serializable {
     /**
      * 用户ID
      */
-    private Integer userId;
+    private Long userId;
 
     /**
      * 订单主键
      */
-    private Integer orderId;
+    private Long orderId;
 
     public Long getId() {
         return id;
@@ -31,19 +31,19 @@ public class Order implements Serializable {
         this.id = id;
     }
 
-    public Integer getUserId() {
+    public Long getUserId() {
         return userId;
     }
 
-    public void setUserId(Integer userId) {
+    public void setUserId(Long userId) {
         this.userId = userId;
     }
 
-    public Integer getOrderId() {
+    public Long getOrderId() {
         return orderId;
     }
 
-    public void setOrderId(Integer orderId) {
+    public void setOrderId(Long orderId) {
         this.orderId = orderId;
     }
 }
