+++ demo-sharding-sphere/src/main/java/com/songjy/sharding/sphere/model/order/OrderItem.java
@@ -18,12 +18,12 @@ public class OrderItem implements Serializable {
     /**
      * 订单ID
      */
-    private Integer orderId;
+    private Long orderId;
 
     /**
      * 用户ID
      */
-    private Integer userId;
+    private Long userId;
 
     /**
      * brandName
@@ -71,19 +71,19 @@ public class OrderItem implements Serializable {
         this.payDiscount = payDiscount;
     }
 
-    public Integer getOrderId() {
+    public Long getOrderId() {
         return orderId;
     }
 
-    public void setOrderId(Integer orderId) {
+    public void setOrderId(Long orderId) {
         this.orderId = orderId;
     }
 
-    public Integer getUserId() {
+    public Long getUserId() {
         return userId;
     }
 
-    public void setUserId(Integer userId) {
+    public void setUserId(Long userId) {
         this.userId = userId;
     }
